Fix the height bug fix half of the animation problem Need to add the fading of animation

+++ GameClient/GameClient/NetUtility.h
@@ -31,11 +31,12 @@ namespace NetUtility
 		message += ":";
 		message += numstr;
 
+		// TODO: refactor with map
 		message += ":";
-		sprintf_s(numstr, "%g", x);
+		sprintf_s(numstr, "%d", (int)x);
 		message += numstr;
 		message += ":";
-		sprintf_s(numstr, "%g", z);
+		sprintf_s(numstr, "%d", (int)z);
 		message += numstr;
 
 		return message;
